ParcialCommit1

+++ sgc/src/main/java/co/edu/eam/sistemasdistribuidos/sgc/core/repositories/BorrowRequestRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BorrowRequestRepository extends JpaRepository<BorrowRequest, Integer> {
+public interface BorrowRequestRepository extends JpaRepository<BorrowRequest, Long> {
 }
